Added google maps service fingerprint and face id.
Switched to elixir server using graphql rest api and absinthe. It is very fast and I like the syntax. :) 
Also updated the user form, and added a calendar booking system. There is testing I started working on for android. to work on it for ios I have to transfer this to a mac. I am considering virtual box :( 
React native and 3 elixir instances will be created